First implementation of Unit tests
 testFindCourseByValidId
Purpose: Validate the functionality of the method when provided with a valid id.

testDeleteCourseByValidId
Purpose: Validate the functionality of the method when provided with a valid id.

1.0.1 - testDeleteCourseByInvalidId
Purpose: Validate the functionality of the method when provided with an invalid id.

testFindCourseByInvalidId
Purpose: Validate the functionality of the method when provided with an invalid id.

testCreateValidCourse()
Purpose: Validate the functionality of the method when creation is successful

Modified EventService class for exception handling

+++ SportCenterManager-Backend/src/main/java/ca/mcgill/ecse321/SportCenterManager/service/EventService.java
@@ -31,13 +31,21 @@ public class EventService {
 
     @Transactional 
     public Course findCourseById(int course_id ){
-        return courseRepo.findCourseById(course_id);
+        Course course = courseRepo.findCourseById(course_id);
+        if (course == null){
+          throw new IllegalArgumentException("There is no course with ID" + course_id + ".");
+        }
+        return course;
     }
 
     @Transactional 
-    public void deleteCourseById(int course_id){
+    public boolean deleteCourseById(int course_id){
         //sessionRepo.deleteByCourse(course_id);
+        if (courseRepo.findCourseById(course_id) == null){
+          throw new IllegalArgumentException("There is no course with ID" + course_id + ".");
+        }
         courseRepo.deleteById(course_id);
+        return true;
 	}
     
         //TODO
